requirements.txt
youtube-transcript-api==0.6.2
google-api-python-client==2.108.0
isodate==0.6.1
openai>=1.55.3
python-dotenv==1.0.0
requests==2.31.0
